+++ georiviere/river/filters.py
@@ -1,9 +1,10 @@
-from django_filters import BooleanFilter, CharFilter, FilterSet, MultipleChoiceFilter
+from django_filters import BooleanFilter, CharFilter, FilterSet, MultipleChoiceFilter, ModelMultipleChoiceFilter
 from django.utils.translation import gettext_lazy as _
 from mapentity.filters import MapEntityFilterSet
 from geotrek.common.filters import OptionalRangeFilter
 from geotrek.zoning.filters import ZoningFilterSet
 
+from georiviere.portal.models import Portal
 from georiviere.river.models import Stream
 from georiviere.watershed.filters import WatershedFilterSet
 
@@ -15,10 +16,20 @@ class StreamFilterSet(WatershedFilterSet, ZoningFilterSet, MapEntityFilterSet):
         label=_('Flow'),
         choices=Stream.FlowChoices.choices
     )
+    portals = ModelMultipleChoiceFilter(
+        method="filter_portal",
+        label=_("Portals"),
+        queryset=Portal.objects.all(),
+    )
 
     class Meta(MapEntityFilterSet.Meta):
         model = Stream
         fields = MapEntityFilterSet.Meta.fields + ['name', 'length', 'flow', 'classification_water_policy']
+
+    def filter_portal(self, qs, name, values):
+        if not values:
+            return qs
+        return qs.prefetch_related('portals').filter(portals__name__in=values)
 
 
 class TopologyFilterSet(FilterSet):