+++ georiviere/river/forms.py
@@ -1,3 +1,5 @@
+from crispy_forms.layout import Div, Field
+
 from django import forms
 
 from geotrek.common.forms import CommonForm
@@ -9,7 +11,19 @@ from georiviere.river.models import Stream, Topology
 
 class StreamForm(CommonForm):
     geom = SnappedLineStringField()
-    geomfields = ['geom', 'source_location']
+    geomfields = ['geom']
+
+    fieldslayout = [
+        Div(
+            "structure",
+            "name",
+            "flow",
+            "main_flow",
+            "data_source",
+            "classification_water_policy",
+            Field('portals', css_class="chzn-select"),
+        )
+    ]
 
     class Meta(CommonForm):
         fields = "__all__"